Update 003: First Commit - Main jump in that uses WinAPI - EngineCore to coordinate the parts - Window class to create a new Window Handle - Renderer class to handle D3D11 creation and use, we'll stick a render pipeline in here later - Folders Set up for better organization
- Should run and show a white window until the window is closed, which should close cleanly.

*Sorry it took a while to push this, I was having trouble getting D3D11 to work
on my older Windows 7 Machine, since you needed special updates with the SDK to
replace the older versions, and the documentation out there can be either highly
detailed or sparce in the need to know info, so finding out my bug with the
D3D11CreateSwapChain was quite a task.

*With that out of the way, I should hopefully be able to work quicker on this after
work. Again, sorry.