Update 004: Viewport Setup - Renderer now has a backBuffer that it holds onto - We setup the Back Buffer and viewport - We clear that BB, render to it(in theory), and then Present it - Cleanup adjusted for new BB, other components for now are temps
- Should run a Black window now

*Using the Swap Chain, we essentially draw our scene in the background before
passing it to the front to be presented on the screen, and then start all over
again with the new back buffer, aka the old front buffer. This is called SwapChaining
and is nearly Identical to OpenGL except for the weird name and room to add more
customized details right away. For now we'll use a basic BB

*This also will be how we'll be rendering through this demo. By using the Renderer
detached from the EngineCore we will be able to fiddle with the Renderer without damage
to other systems. Not only that, but it will keep our two seperate systems apart and easier
to compartmentalize for when we start doing weirder techniques.